+++ tests/testthat/test-module-prereg.R
@@ -98,11 +98,36 @@ test_that("multiple papers", {
   expect_equal(nrow(mo$table), 2)
   expect_setequal(mo$table$template_name,
                   c("OSF Preregistration", "AsPredicted"))
+  expect_contains(names(mo$table), "paper_id")
   expect_setequal(mo$table$id, c(guid1, guid2))
   ids <- paper_id(paper)$paper_id
   expect_setequal(mo$table$paper_id, ids)
   expect_setequal(mo$summary_table$paper_id, ids)
   expect_setequal(mo$summary_table$preregistration, c(1,1))
+})
+
+test_that("combine >10 OSF registrations", {
+  # https://github.com/scienceverse/metacheck/issues/262
+
+  # with >10 reg, osf_get_all_pages will paginate
+  # and the different reg forms often return incompatible DFs
+  reg_urls <- c("https://osf.io/bdvxs",
+               "https://osf.io/wrh4x",
+               "https://osf.io/trwb4",
+               "https://osf.io/z2bsa",
+               "https://osf.io/jez3g",
+               "https://osf.io/9bg3z",
+               "https://osf.io/7qcxa",
+               "https://osf.io/a6y7r",
+               "https://osf.io/4v3sg",
+               "https://osf.io/hwu9x",
+               "https://osf.io/yab8q")
+  paper <- test_paper(url = reg_urls)
+  mo <- module_run(paper, "prereg_check")
+
+  expect_equal(mo$summary_table$preregistration, 11)
+  expect_equal(nrow(mo$table), 11)
+  expect_contains(names(mo$table), "paper_id")
 })
 
 httptest2::stop_mocking()